refactor /user/ pages to the newer code/template split
This takes the existing `/user/` page and splits it up to put the html in
templates and a chunk of the logic in the `User` class with only the
display login in `/user/index.php`.

Ideally this should be split into multiple pages:
 * view user
 * edit user
 * join

and move the other user editing in to the admin interface but that's one
more step than I want to take.